+++ discopy/messages.py
@@ -10,7 +10,6 @@ NOT_PARALLEL = "Expected parallel arrows, got {} and {} instead."
 NOT_GLOBULAR = "Expected a globular box, got boundaries {} -> {} and {} -> {}."
 NOT_ATOMIC = "Expected {} of length 1, got length {} instead."
 NOT_CONNECTED = "{} is not boundary-connected."
-NOT_PLANAR = "{} is not planar."
 NOT_TRACEABLE = "Cannot trace {} with {}."
 NOT_ADJOINT = "{} and {} are not adjoints."
 NOT_RIGID_ADJOINT = "{} is not the left adjoint of {}, maybe you meant to use"\